+++ app/api/transaction/route.ts
@@ -1,11 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { GetRealEstateTransactionUseCase } from '@be/applications/realEstate/usecases/GetRealEstateTransactionUseCase';
-import { GetRealEstateTransactionRequest } from '@be/applications/realEstate/dtos/GetRealEstateTransactionRequest';
-import { GetRealEstateTransactionResponse } from '@be/applications/realEstate/dtos/GetRealEstateTransactionResponse';
+import { GetRealEstateTransactionUseCase } from '@be/applications/transaction/usecases/GetRealEstateTransactionUseCase';
+import { GetRealEstateTransactionRequest } from '@be/applications/transaction/dtos/GetRealEstateTransactionRequest';
+import { GetRealEstateTransactionResponse } from '@be/applications/transaction/dtos/GetRealEstateTransactionResponse';
 
 /**
  * 실거래가 조회 API
- * GET /api/real-estate/search/transaction-price
+ * GET /api/transaction/transaction-price
  */
 export async function GET(request: NextRequest) {
   try {
