+++ backend/infrastructure/repository/RealEstateTransactionRepository.ts
@@ -1,6 +1,5 @@
-import axios from 'axios';
-import { GetRealEstateTransactionRequest } from '@be/applications/realEstate/dtos/GetRealEstateTransactionRequest';
-import { GetRealEstateTransactionResponse } from '@be/applications/realEstate/dtos/GetRealEstateTransactionResponse';
+import { GetRealEstateTransactionRequest } from '@be/applications/transaction/dtos/GetRealEstateTransactionRequest';
+import { GetRealEstateTransactionResponse } from '@be/applications/transaction/dtos/GetRealEstateTransactionResponse';
 import PublicDataAxiosInstance from '@utils/axiosInstance';
 import { parseXmlResponse } from '@utils/xmlParser';
